tests(remote): Use 'BRANCH' var with 'pytest' default value
When tests use 'main' branch, local pytest runs can differ from pipeline runs.

For example: There is no git repo when pytest runs in the pipeline.
But locally it runs inside the git-remote-s3 repo.
If mocking of a git command wasn't done properly, local tests pass because a git repo with a main branch exist.
They then fail in the pipeline with an error like: 'fatal: not a valid object name: refs/heads/main'

By moving away from a branch name that exists locally, the same behavior as in a pipeline run is enforced and problems easier to discover.